+++ ndd/nsb.py
@@ -2,191 +2,43 @@
 # Copyright (C) 2016,2017 Simone Marsili
 # All rights reserved.
 # License: BSD 3 clause
-from __future__ import (absolute_import, division,
-                        print_function, unicode_literals)
+"""Functions module."""
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
 from builtins import (  # pylint: disable=redefined-builtin, unused-import
-    bytes, dict, int, list, object, range, str,
-    ascii, chr, hex, input, next, oct, open,
-    pow, round, super,
-    filter, map, zip)
+    bytes, dict, int, list, object, range, str, ascii, chr, hex, input, next,
+    oct, open, pow, round, super, filter, map, zip)
+import numpy
+import ndd
+from ndd.estimators import Entropy, JSDivergence
 
 __copyright__ = "Copyright (C) 2016,2017 Simone Marsili"
 __license__ = "BSD 3 clause"
 __author__ = "Simone Marsili (simomarsili@gmail.com)"
 __all__ = ['entropy', 'histogram', 'from_data', 'nbins']
 
-import numpy
-import ndd
-import ndd._nsb
 
-
-class BaseEstimator(object):
-    def __init__(self):
-        self.estimate = None
-        self.std = None
-
-    def _check(self):
-        # check input data
-        raise NotImplementedError
-
-    def fit(self):
-        # set self.estimate, self.std
-        raise NotImplementedError
-
-    def __call__(self, *args, **kwargs):
-        """Return estimate from input data. Delegate to fit."""
-        self.fit(*args, **kwargs)
-        return self.estimate
-
-
-class Entropy(BaseEstimator):
-    def __init__(self, alpha=None, plugin=False):
-        super().__init__()
-
-        # check alpha value
-        if alpha:
-            try:
-                alpha = numpy.float64(alpha)
-            except ValueError:
-                raise
-            if alpha <= 0:
-                raise ValueError("alpha <= 0")
-        self.alpha = alpha
-
-        # set estimator
-        if plugin:
-            if alpha is None:
-                self.estimator = self._plugin
-            else:
-                self.estimator = lambda counts, k: self._pseudocounts(
-                    counts, k, self.alpha)
-        else:
-            if alpha is None:
-                self.estimator = self._nsb
-            else:
-                self.estimator = lambda counts, k: self._ww(
-                    counts, k, self.alpha)
-
-    @staticmethod
-    def _plugin(counts, k):
-        return ndd._nsb.plugin(counts, k)
-
-    @staticmethod
-    def _pseudocounts(counts, k, alpha):
-        return ndd._nsb.pseudo(counts, k, alpha)
-
-    @staticmethod
-    def _ww(counts, k, alpha):
-        return ndd._nsb.dirichlet(counts, k, alpha)
-
-    @staticmethod
-    def _nsb(counts, k):
-        return ndd._nsb.nsb(counts, k)
-
-    def _check(self, counts, k):
-        counts = self._check_counts(a=counts)
-        k = self._check_k(k=k, n_bins=len(counts))
-        return counts, k
-
-    @staticmethod
-    def _check_counts(a):
-        try:
-            a = numpy.asarray(a, dtype=numpy.int32)
-        except ValueError:
-            raise
-        if numpy.any(a < 0):
-            raise ValueError("Frequency counts can't be negative")
-        return a.flatten()
-
-    @staticmethod
-    def _check_k(k, n_bins):
-        """
-        if k is None, set k = number of bins
-        if k is an integer, just check
-        ik an array set k = prod(k)
-        """
-        MAX_LOGK = 150 * numpy.log(2)
-
-        if k is None:
-            # set k to the number of observed bins
-            k = numpy.float64(n_bins)
-        else:
-            try:
-                k = numpy.float64(k)
-            except ValueError:
-                raise
-            if k.ndim:
-                # if k is a sequence, set k = prod(k)
-                if k.ndim > 1:
-                    raise ValueError('k must be a scalar or 1D array')
-                logk = numpy.sum(numpy.log(x) for x in k)
-                if logk > MAX_LOGK:
-                    # too large a number; backoff to n_bins?
-                    # TODO: log warning
-                    raise ValueError('k (%r) larger than %r' %
-                                     (numpy.exp(logk), numpy.exp(MAX_LOGK)))
-                else:
-                    k = numpy.prod(k)
-            else:
-                # if a scalar check size
-                if numpy.log(k) > MAX_LOGK:
-                    raise ValueError('k (%r) larger than %r' %
-                                     (k, numpy.exp(MAX_LOGK)))
-            # consistency checks
-            if k < n_bins:
-                raise ValueError("k (%s) is smaller than the number of bins"
-                                 "(%s)" % (k, n_bins))
-            if not k.is_integer():
-                raise ValueError("k (%s) should be a whole number." % k)
-        return k
-
-    def fit(self, counts, k=None):
-        """
-        counts : array_like
-            The number of occurrences of a set of bins.
-
-        k : int, optional
-            Number of bins. k >= len(counts).
-            Float values are valid input for whole numbers (e.g. k=1.e3).
-            Defaults to len(counts).
-
-        """
-        counts, k = self._check(counts, k)
-        if k == 1:  # single bin
-            self.estimate = self.std = 0.0
-        else:
-            result = self.estimator(counts, k)
-            if isinstance(result, tuple):
-                self.estimate, self.std = result
-            else:
-                self.estimate = result
-
-
-def entropy(counts, k=None, alpha=None, return_std=False, plugin=False):
+def entropy(pk, k=None, alpha=None, plugin=False, return_std=False):
     """
-    Return a Bayesian estimate of the entropy of an unknown discrete
-    distribution from an input array of counts.
+    Return a Bayesian estimate S' of the entropy of an unknown discrete
+    distribution from an input array of counts pk.
 
     Parameters
     ----------
 
-    counts : array_like
+    pk : array-like
         The number of occurrences of a set of bins.
-
     k : int or array-like, optional
-        Number of bins. k >= len(counts).
+        Number of bins; k >= len(pk).
         A float value is a valid input for whole numbers (e.g. k=1.e3).
         If an array, set k = numpy.prod(k).
-        Defaults to len(counts).
-
+        Defaults to len(pk).
     alpha : float, optional
-        If alpha is passed, use a single Dirichlet prior with concentration
+        If alpha is not None, use a single Dirichlet prior with concentration
         parameter alpha (fixed alpha estimator). alpha > 0.0.
-
     return_std : boolean, optional
         If True, also return an approximated value for the standard deviation
         over the entropy posterior.
-
     plugin : boolean, optional
         If True, return a 'plugin' estimate of the entropy. The discrete
         distribution is estimated from the empirical frequencies over bins
@@ -198,26 +50,71 @@ def entropy(counts, k=None, alpha=None, return_std=False, plugin=False):
     -------
     entropy : float
         Entropy estimate.
-
     std : float, optional
         Uncertainty in the entropy estimate
-        (approximates the standard deviation over the entropy posterior).
+        (approximated standard deviation over the entropy posterior).
         Only provided if `return_std` is True.
 
     """
 
-    estimator = Entropy(alpha, plugin)
-    estimator.fit(counts, k)
+    # pk is an array of counts
+    estimator = Entropy(alpha, plugin).fit(pk, k)
+    S, err = estimator.estimate_, estimator.err_
 
-    if return_std:
-        result = estimator.estimate, estimator.std
-    else:
-        result = estimator.estimate
-
-    if numpy.any(numpy.isnan(numpy.squeeze(result))):
+    if numpy.isnan(S) or (err is not None and numpy.isnan(err)):
         raise FloatingPointError("NaN value")
 
-    return result
+    if return_std:
+        return S, err
+    else:
+        return S
+
+
+def jensen_shannon_divergence(pk, k=None, alpha=None, plugin=False):
+    """
+    Estimate the Jensen-Shannon divergence from a matrix of counts.
+
+    Return an estimate of the Jensen-Shannon divergence between
+    n unknown discrete distributions from a n-by=p input array of
+    counts. The estimate is computed as a combination of single Bayesian
+    entropy estimates, and is bounded by ln(n). The combination is weighted by
+    the total number of samples for each distribution, see:
+    https://en.wikipedia.org/wiki/Jensen-Shannon_divergence
+
+    Parameters
+    ----------
+
+    pk : array-like, shape (n_distributions, p)
+        Different rows correspond to counts from different distributions with
+        the same discrete sample space.
+    k : int or array-like, optional
+        Number of bins; k >= len(pk).
+        A float value is a valid input for whole numbers (e.g. k=1.e3).
+        If an array, set k = numpy.prod(k).
+        Defaults to pk.shape[1].
+    alpha : float, optional
+        If not None, the entropy estimator uses a single Dirichlet prior with
+        concentration parameter alpha (fixed alpha estimator). alpha > 0.0.
+    plugin : boolean, optional
+        If True, use a 'plugin' estimator for the entropy.
+        If alpha is passed in combination with plugin == True, add alpha
+        pseudoconts to the frequency counts in the plugin estimate.
+
+    Returns
+    -------
+    float
+        JS divergence estimate.
+
+    """
+
+    # pk is an array of counts
+    estimator = JSDivergence(alpha, plugin).fit(pk, k)
+    js_div = estimator.estimate_
+
+    if numpy.isnan(js_div):
+        raise FloatingPointError("NaN value")
+
+    return js_div
 
 
 def nbins(data):
@@ -226,7 +123,7 @@ def nbins(data):
     the num. of unique elements for each variable.
     """
     # reshape as a p-by-n array
-    data = ndd.nsb._as_data_array(data)
+    data = ndd.nsb.as_data_array(data)
     return [len(numpy.unique(v)) for v in data]
 
 
@@ -251,11 +148,11 @@ def histogram(data, axis=0, r=0):
     """
     from itertools import combinations
     # reshape as a p-by-n array
-    data = ndd.nsb._as_data_array(data, axis=axis)
-    p, n = data.shape
+    data = ndd.nsb.as_data_array(data, axis=axis)
+    p = data.shape[0]
     if r > p:
-        raise ValueError('r (%r) is larger than the number of variables (%r)'
-                         % (r, p))
+        raise ValueError(
+            'r (%r) is larger than the number of variables (%r)' % (r, p))
     if r == 0:
         # statistics for the p-dimensional variable
         _, counts = numpy.unique(data, return_counts=True, axis=1)
@@ -264,7 +161,7 @@ def histogram(data, axis=0, r=0):
         return (ndd.histogram(d, axis=1) for d in combinations(data, r=r))
 
 
-def _as_data_array(ar, axis=0):
+def as_data_array(ar, axis=0):
     """
     For a 2D n-by-p data array, transpose it.
     For a generic ndarray, flatten the subarrays indexed by axis 0
@@ -317,8 +214,8 @@ def from_data(ar, ks=None, axis=0, r=0):
     """
     from itertools import combinations
 
-    ar = _as_data_array(ar, axis=axis)
-    p, n = ar.shape
+    ar = as_data_array(ar, axis=axis)
+    p = ar.shape[0]
 
     if ks is None:
         ks = numpy.array([len(numpy.unique(v)) for v in ar])
@@ -340,5 +237,6 @@ def from_data(ar, ks=None, axis=0, r=0):
         counts_combinations = histogram(ar, axis=1, r=r)
         alphabet_size_combinations = (numpy.prod(x)
                                       for x in combinations(ks, r=r))
-        return (entropy_estimator(c, k=k) for c, k in
-                zip(counts_combinations, alphabet_size_combinations))
+        return (
+            entropy_estimator(c, k=k)
+            for c, k in zip(counts_combinations, alphabet_size_combinations))